requirements.txt
langchain==0.1.0
langchain_core==0.1.8

numpy==1.25.2
python-dotenv==1.0.0

scikit_learn==1.3.2
streamlit==1.29.0
text2vec==1.2.9

streamlit